Test compression with invalid content type for all webservers
Closes gh-18338
